[Divide by 0 Error] normalize check migrate to c++
