--- python/paddle/fluid/tests/unittests/test_nn_functional_norm_op.py
@@ -1,37 +1,0 @@
-#  Copyright (c) 2023 PaddlePaddle Authors. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#    http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-import unittest
-
-import numpy as np
-
-import paddle
-
-
-class TestSparseEmbeddingAPIError(unittest.TestCase):
-    def test_errors(self):
-        with paddle.fluid.dygraph.guard():
-            # The size of input in sparse_embedding should not be 0.
-            def test_0_size():
-                array = np.array([], dtype=np.float32)
-                x = paddle.to_tensor(
-                    np.reshape(array, [1, 1, 0]), dtype='float32'
-                )
-                paddle.nn.functional.normalize(x)
-
-            self.assertRaises(ValueError, test_0_size)
-
-
-if __name__ == '__main__':
-    unittest.main()
