+++ python/paddle/fluid/tests/unittests/test_normalize.py
@@ -97,6 +97,18 @@ class TestNNFunctionalNormalize(unittest.TestCase):
         with fluid.program_guard(fluid.Program()):
             self.run_static(use_gpu=True)
 
+    def test_errors(self):
+        with fluid.dygraph.guard():
+            # The size of input in Normalize should not be 0.
+            def test_0_size():
+                array = np.array([], dtype=np.float32)
+                x = paddle.to_tensor(
+                    np.reshape(array, [1, 1, 0]), dtype='float32'
+                )
+                paddle.nn.functional.normalize(x)
+
+            self.assertRaises(ValueError, test_0_size)
+
 
 if __name__ == "__main__":
     unittest.main()
