+++ python/paddle/nn/functional/norm.py
@@ -78,9 +78,6 @@ def normalize(x, p=2, axis=1, epsilon=1e-12, name=None):
             #         [1.        , 0.97014254, 0.92847669]])
     """
 
-    if x.size == 0:
-        raise ValueError("input size should not be 0")
-
     if in_dygraph_mode():
         eps = fluid.dygraph.base.to_variable([epsilon], dtype=x.dtype)
         out = _C_ops.p_norm(x, float(p), axis, epsilon, True, False)
